Tested windows serial
Needs linux version and added to cmake. Both network connections could
use better error handling if needed.

+++ CommProto/src/network/Serial.cpp
@@ -66,7 +66,7 @@ bool Serial::initWindows(std::string comPort, uint16_t baudrate)
 		return false;
 	}
 
-	dcbSerialParams.BaudRate = baudrate;
+	dcbSerialParams.BaudRate = (DWORD)baudrate;
 	dcbSerialParams.ByteSize = 8;
 	dcbSerialParams.StopBits = ONESTOPBIT;
 	dcbSerialParams.Parity = NOPARITY;
